refactor mobility calculations in submodule
